update: double型に変更を行った。

+++ prg15/prog6.c
@@ -44,14 +44,14 @@ int main(int argc, const char* argv[]){
 
     // 推定部分
 
-    int MSE = 0;
+    double MSE = 0;
     for(int i = 0; i < inputIndex; i++){
 
-        int estimatedDistance = estimationFunction(trialList[i].speed);
+        double estimatedDistance = estimationFunction(trialList[i].speed);
 
-        int error = estimatedDistance - trialList[i].distance;
+        double error = estimatedDistance - trialList[i].distance;
 
-        printf("speed: %d, distance: %d, 推定: %d, 誤差: %d\n", trialList[i].speed, trialList[i].distance, estimatedDistance, error);
+        printf("speed: %d, distance: %d, 推定: %f, 誤差: %f\n", trialList[i].speed, trialList[i].distance, estimatedDistance, error);
 
         //Mean Square Error
         MSE += error * error;
@@ -59,7 +59,7 @@ int main(int argc, const char* argv[]){
 
     MSE = MSE / inputIndex;
 
-    printf("MSE: %d\n", MSE);
+    printf("MSE: %f\n", MSE);
